Lift CFG to an IL
Not quite complete, variable declarations/assignments aren't done right yet. But almost there.

+++ SmxDecompiler/smx-file.h
@@ -60,7 +60,7 @@ struct SmxESField
 struct SmxEnumStruct
 {
 	const char* name;
-	int num_fields;
+	size_t num_fields;
 	SmxESField* fields;
 	uint32_t size;
 };
